ZM - Adding Snowball image and click code
I added a snowball image to the resources (Needs a little bit of work becuase there is still a little bit of a background on it) I also added some click code.  I commented some of it out.  All it is doing right now is where you click the mouse John snow will teleport there.  This was a test to figure out how to get the snowball to move the way that we talked yesterday