ZM - Added John Snow throwing a snowball
I added the code to have John Snow throw a snowball.  The snowball will
start where John Snow is and it will end up where the user clicked.
This is a mouse down event on the form.

This is okay but not the best I would like to see the snowball move
across the form better than what it does right now.